move reporting provider models to their own files.
