+++ koku/reporting/provider/ocp/__init__.py
@@ -1,0 +1,1 @@
+# noqa
